Food2 go trenutna fix

+++ FOOD2GO/app/src/main/java/hr/foi/air/food2go/controller/dataLoaders/WsDataLoader.java
@@ -92,7 +92,7 @@ public class WsDataLoader {
         webServiceCaller.DodajArtiklNaRacun(artikl,racun);
     }
 
-    public void DodajCijenuNaRacun(Racun racun, int cijena) {
+    public void DodajCijenuNaRacun(Racun racun, float cijena) {
         this.dataLoadedListener = dataLoadedListener;
         Log.i("AIR_WSDataLoader", String.valueOf(cijena));
         webServiceCaller.DodajCijenuRacunu(racun, cijena);
